ISS-226 # JUnit5 Jupiter Load Extension - new module
