--- junit5-testing/src/main/java/org/jsmart/zerocode/zerocodejavaexec/httpclient/CustomHttpClient.java
@@ -1,11 +1,0 @@
-package org.jsmart.zerocode.zerocodejavaexec.httpclient;
-
-import org.jsmart.zerocode.core.httpclient.BasicHttpClient;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class CustomHttpClient extends BasicHttpClient {
-    private static final Logger LOGGER = LoggerFactory.getLogger(CustomHttpClient.class);
-    //TODO - With security headers
-}
-
