+++ junit5-testing/src/test/java/org/jsmart/zerocode/jupiter/load/JupiterLoadProcessorTest.java
@@ -1,0 +1,5 @@
+package org.jsmart.zerocode.jupiter.load;
+
+class JupiterLoadProcessorTest {
+
+}
