+++ junit5-testing/src/test/java/org/jsmart/zerocode/tests/loadjupiter/JUnit5LoadTest.java
@@ -3,12 +3,12 @@ package org.jsmart.zerocode.tests.loadjupiter;
 import org.jsmart.zerocode.core.domain.LoadWith;
 import org.jsmart.zerocode.core.domain.TestMapping;
 import org.jsmart.zerocode.core.domain.TestMappings;
-import org.jsmart.zerocode.jupiter.ZeroCodeJupiterLoad;
+import org.jsmart.zerocode.jupiter.extension.ZeroCodeParallelExtension;
 import org.jsmart.zerocode.tests.jupiter.JUnit5Test;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-@ExtendWith({ZeroCodeJupiterLoad.class})
+@ExtendWith({ZeroCodeParallelExtension.class})
 public class JUnit5LoadTest {
 
     @Test
